// backend/src/modules/production-orders/dto/update-production-order-disintegration-check.dto.ts
export class UpdateProductionOrderDisintegrationCheckDto {
  requirement?: string | null;
  dosage_form_stage?: string | null;
  unit_1_passed?: boolean | number | string | null;
  unit_2_passed?: boolean | number | string | null;
  unit_3_passed?: boolean | number | string | null;
  unit_4_passed?: boolean | number | string | null;
  unit_5_passed?: boolean | number | string | null;
  unit_6_passed?: boolean | number | string | null;
}